+++ packages/asc-ui/src/components/Footer/FooterContent/FooterContentStyle.ts
@@ -10,7 +10,6 @@ export default styled.div<Props>`
   display: flex;
   flex-direction: column;
   margin-left: ${({ indent }) => (indent ? '28px' : '0')};
-  max-width: 300px;
   ${showHide}
 
   ${ParagraphStyle} {